requirements.txt
flask==1.1.2
flask-sqlalchemy==2.4.4
python-dotenv==0.14.0
dataclasses==0.6
psycopg2==2.8.6
SQLAlchemy~=1.3.19
